+++ Tests/Template.Tests.ps1
@@ -1,18 +1,20 @@
 $credentials = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList $Global:OMEUsername, $(ConvertTo-SecureString -Force -AsPlainText $Global:OMEPassword)
 Connect-OMEServer -Name $Global:OMEServer -Credentials $credentials -IgnoreCertificateWarning
-Describe "Template" {
+Describe "Template Tests" {
     BeforeAll {
         $Script:DeviceServiceTag1 = "37KP0Q2"
         $Script:DeviceServiceTag2 = "C86F0Q2"
         $Script:DeviceServiceTag3 = "GV6V673"
         $Script:ConfigurationBaselineName = "TestConfigurationBaseline_$((Get-Date).ToString('yyyyMMddHHmmss'))"
+        $Script:DeploymentTemplateNameFromString = "TestDeploymentTemplate_FromString_$((Get-Date).ToString('yyyyMMddHHmmss'))"
+        $Script:ConfigurationTemplateNameFromString = "TestConfigurationTemplate_FromString_$((Get-Date).ToString('yyyyMMddHHmmss'))"
     }
     Context "General" {
         It "Should return ALL templates" {
             Get-OMETemplate | Measure-Object | Select-Object -ExpandProperty Count | Should -BeGreaterThan 2
         }
     }
-    Context "Deployment" {
+    Context "Deployment" -Tag "Deployment" {
         It ("Should create a deployment template from source device and return JobId"){
             $TemplateName = "TestDeploymentTemplate_FromDevice_$((Get-Date).ToString('yyyyMMddHHmmss'))"
             New-OMETemplateFromDevice -Name $TemplateName -Device $($Script:DeviceServiceTag1 | Get-OMEDevice -FilterBy "ServiceTag") -Component "All" | Should -BeGreaterThan 0
@@ -34,10 +36,9 @@ Describe "Template" {
                 </Component>
             </SystemConfiguration>
             '
-            $TemplateNameFromString = "TestDeploymentTemplate_FromString_$((Get-Date).ToString('yyyyMMddHHmmss'))"
-            New-OMETemplateFromFile -Name $TemplateNameFromString -Content $xml -Wait
-            $TemplateNameFromString | Get-OMETemplate -FilterBy "Name" | Select-Object -ExpandProperty Name | Should -Be $TemplateNameFromString
-            $TemplateNameFromString | Get-OMETemplate -FilterBy "Name" | Select-Object -ExpandProperty ViewTypeId | Should -Be 2
+            New-OMETemplateFromFile -Name $Script:DeploymentTemplateNameFromString -Content $xml -Wait
+            $Script:DeploymentTemplateNameFromString | Get-OMETemplate -FilterBy "Name" | Select-Object -ExpandProperty Name | Should -Be $Script:DeploymentTemplateNameFromString
+            $Script:DeploymentTemplateNameFromString | Get-OMETemplate -FilterBy "Name" | Select-Object -ExpandProperty ViewTypeId | Should -Be 2
         }
 
         It "Should return all deployment templates" {
@@ -49,15 +50,15 @@ Describe "Template" {
         }
 
         It "Should deploy template to device" {
-            $template = $("TestDeploymentTemplate_FromString" | Get-OMETemplate | Select-Object -first 1)
+            $template = $($Script:DeploymentTemplateNameFromString | Get-OMETemplate)
             $devices = $Script:DeviceServiceTag1, $Script:DeviceServiceTag2, $Script:DeviceServiceTag3 | Get-OMEDevice
-            Invoke-OMETemplateDeploy -Template $template -Devices $devices -Wait | Should -BeIn @("Completed", "Warning")
+            Invoke-OMETemplateDeploy -Template $template -Devices $devices -Wait | Should -Match "Completed.*"
         }
     }
-    Context "Configuration" {
+    Context "Configuration" -Tag "Configuration" {
         It ("Should create a compliance template from source device and return JobId"){
             $TemplateName = "TestComplianceTemplate_FromDevice_$((Get-Date).ToString('yyyyMMddHHmmss'))"
-            New-OMETemplateFromDevice -Name $TemplateName -TemplateType "Configuration" -Device $($Script:DeviceServiceTag1 | Get-OMEDevice -FilterBy "ServiceTag") -Component "All" | Should -BeGreaterThan 0
+            New-OMETemplateFromDevice -Name $TemplateName -TemplateType "Configuration" -Device $($Script:DeviceServiceTag2 | Get-OMEDevice -FilterBy "ServiceTag") -Component "All" | Should -BeGreaterThan 0
         }
 
         It ("Should create a new compliance template from XML string located in a file") {
@@ -76,10 +77,9 @@ Describe "Template" {
                 </Component>
             </SystemConfiguration>
             '
-            $TemplateNameFromString = "TestComplianceTemplate_FromString_$((Get-Date).ToString('yyyyMMddHHmmss'))"
-            New-OMETemplateFromFile -Name $TemplateNameFromString -TemplateType "Configuration" -Content $xml -Wait
-            $TemplateNameFromString | Get-OMETemplate -FilterBy "Name" | Select-Object -ExpandProperty Name | Should -Be $TemplateNameFromString
-            $TemplateNameFromString | Get-OMETemplate -FilterBy "Name" | Select-Object -ExpandProperty ViewTypeId | Should -Be 1
+            New-OMETemplateFromFile -Name $Script:ConfigurationTemplateNameFromString -TemplateType "Configuration" -Content $xml -Wait
+            $Script:ConfigurationTemplateNameFromString | Get-OMETemplate -FilterBy "Name" | Select-Object -ExpandProperty Name | Should -Be $Script:ConfigurationTemplateNameFromString
+            $Script:ConfigurationTemplateNameFromString | Get-OMETemplate -FilterBy "Name" | Select-Object -ExpandProperty ViewTypeId | Should -Be 1
         }
 
         It "Should return all compliance templates" {
@@ -87,9 +87,9 @@ Describe "Template" {
         }
 
         It "Should create configuration compliance baseline" {
-            $devices = $Script:DeviceServiceTag1 | Get-OMEDevice -FilterBy "ServiceTag"
-            $template = "TestComplianceTemplate_FromString" | Get-OMETemplate | Select-Object -first 1
-            New-OMEConfigurationBaseline -Name $Script:ConfigurationBaselineName -Template $template -Devices $devices -Wait | Should -Be "Completed"
+            $devices = $Script:DeviceServiceTag3 | Get-OMEDevice -FilterBy "ServiceTag"
+            $template = $Script:ConfigurationTemplateNameFromString | Get-OMETemplate
+            New-OMEConfigurationBaseline -Name $Script:ConfigurationBaselineName -Template $template -Devices $devices -Wait | Should -Match "Completed.*"
         }
 
         It "Should return configuration compliance baseline" {
@@ -98,28 +98,28 @@ Describe "Template" {
 
         It "Should update configuration compliance on device" {
             $baseline = $Script:ConfigurationBaselineName | Get-OMEConfigurationBaseline
-            $devices = $Script:DeviceServiceTag1 | Get-OMEDevice -FilterBy "ServiceTag"
-            Update-OMEConfiguration -Name "Make Compliant $((Get-Date).ToString('yyyyMMddHHmmss'))" -Baseline $baseline -DeviceFilter $devices -Wait | Should -Be "Completed"
+            $devices = $Script:DeviceServiceTag3 | Get-OMEDevice -FilterBy "ServiceTag"
+            Update-OMEConfiguration -Name "Make Compliant $((Get-Date).ToString('yyyyMMddHHmmss'))" -Baseline $baseline -DeviceFilter $devices -Wait | Should -Match "Completed.*"
         }
 
-        It "Should check configuration compliance for baseline" {
+        It "Should check configuration compliance for baseline" -Tag "CheckConfigurationCompliance" {
             $baseline = $Script:ConfigurationBaselineName | Get-OMEConfigurationBaseline
-            $baseline  | Invoke-OMEConfigurationBaselineRefresh -Wait | Should -Be "Completed"
+            $baseline  | Invoke-OMEConfigurationBaselineRefresh | Should -BeOfType [int]
         }
     }
-    Context "Profile" {
+    Context "Profile" -Tag "Profile" {
         It "Should unassign profile by device" {
             $devices = $Script:DeviceServiceTag1 | Get-OMEDevice
-            Invoke-OMEProfileUnassign -Device $devices -Wait | Should -Be "Completed"
+            Invoke-OMEProfileUnassign -Device $devices | Should -BeOfType [int]
         }
 
         It "Should unassign profile by profile name" {
-            Invoke-OMEProfileUnassign -ProfileName "00002" -Wait | Should -Be "Completed"
+            Invoke-OMEProfileUnassign -ProfileName "00002" | Should -BeOfType [int]
         }
 
         It "Should unassign profile by template" {
-            $template = "TestDeploymentTemplate_FromString" | Get-OMETemplate | Select-Object -first 1
-            Invoke-OMEProfileUnassign -Template $template -Wait | Should -Be "Completed"
+            $template = $Script:DeploymentTemplateNameFromString | Get-OMETemplate
+            Invoke-OMEProfileUnassign -Template $template | Should -BeOfType [int]
         }
     }
 }